long

+++ modules/apps/foundation/portal-cache/portal-cache-cache2k/src/main/java/com/liferay/portal/cache/cache2k/internal/Cache2kPortalCache.java
@@ -68,14 +68,16 @@ public class Cache2kPortalCache<K extends Serializable, V>
 		cache.put(key, value);
 
 		if (timeToLive != DEFAULT_TIME_TO_LIVE) {
-			cache.expireAt(key, System.currentTimeMillis() + timeToLive * 1000);
+			cache.expireAt(
+				key, System.currentTimeMillis() + (long)timeToLive * 1000);
 		}
 	}
 
 	@Override
 	protected V doPutIfAbsent(K key, V value, int timeToLive) {
 		if (cache.putIfAbsent(key, value)) {
-			cache.expireAt(key, System.currentTimeMillis() + timeToLive * 1000);
+			cache.expireAt(
+				key, System.currentTimeMillis() + (long)timeToLive * 1000);
 
 			return null;
 		}
@@ -100,7 +102,8 @@ public class Cache2kPortalCache<K extends Serializable, V>
 
 			cache.put(key, value);
 
-			cache.expireAt(key, System.currentTimeMillis() + timeToLive * 1000);
+			cache.expireAt(
+				key, System.currentTimeMillis() + (long)timeToLive * 1000);
 
 			return oldValue;
 		}
@@ -111,7 +114,8 @@ public class Cache2kPortalCache<K extends Serializable, V>
 	@Override
 	protected boolean doReplace(K key, V oldValue, V newValue, int timeToLive) {
 		if (cache.replaceIfEquals(key, oldValue, newValue)) {
-			cache.expireAt(key, System.currentTimeMillis() + timeToLive * 1000);
+			cache.expireAt(
+				key, System.currentTimeMillis() + (long)timeToLive * 1000);
 
 			return true;
 		}
